% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cier-package.R
\docType{package}
\name{cier-package}
\alias{cier}
\alias{cier-package}
\title{cier: Detect Careless and Insufficient Effort Responding in Survey Data}
\description{
A lean, fast toolkit for detecting careless and insufficient effort responding ('C/IER') in survey data using indirect response-pattern indices (longstring, intra-individual response variability, personal reliability, psychometric synonyms and antonyms, Mahalanobis distance, person-total correlation, and nonparametric person-fit statistics), following the indirect-index battery evaluated by Goldammer et al. (2024) \doi{10.3758/s13428-024-02484-3}. Flags respondents through an auditable, false-positive-rate-controlled layer with transparent, citable defaults.
}
\author{
\strong{Maintainer}: Markus Steiner \email{markus.d.steiner@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Markus Steiner \email{markus.d.steiner@gmail.com} [copyright holder]
}

}
\keyword{internal}
