% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cier-screen.R
\name{print.cier_screen}
\alias{print.cier_screen}
\title{Print a cier screen}
\usage{
\method{print}{cier_screen}(x, ...)
}
\arguments{
\item{x}{A \code{cier_screen} object, as returned by \code{\link[=cier_screen]{cier_screen()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a cier screen
}
